+++ src/games/planetwars/game_examples/standard_game.json
