+++ src/games/planetwars/visualizer/index.html
@@ -9,6 +9,8 @@
 
 <body>
   <div id="visualizer">
+
+    <!-- -->
     <svg id="score"></svg>
 
     <!-- Hide/unhide button for scoreboard -->
