+++ src/games/planetwars/visualizer/src/assets/images/earth.png
