+++ src/games/planetwars/visualizer/src/assets/images/logo.png
