+++ src/games/planetwars/visualizer/src/assets/images/mars.png
