+++ src/games/planetwars/visualizer/src/assets/images/marsss.png
