+++ src/games/planetwars/visualizer/src/assets/images/moon.png
