+++ src/games/planetwars/visualizer/src/assets/images/red.png
