+++ src/games/planetwars/visualizer/src/assets/images/rocket.png
