--- src/games/planetwars/visualizer/style.css
@@ -1,273 +1,0 @@
-@import url('./node_modules/font-awesome/css/font-awesome.min.css');
-
-/* Keep this at the top so it keeps having priority */
-.invisible {
-  display: none !important;
-}
-
-body {
-  margin: 0;
-}
-
-#box {
-  height: 100vh;
-  width: 100vw;
-}
-
-#blockly {
-  width: 100%;
-  height: 100%;
-}
-
-#visualizer {
-  height: 100%;
-}
-
-#game {
-  width: 100%;
-  height: 99.7vh;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-#score {
-  width: 20%;
-  height: 50vh;
-  position: absolute;
-  right: 1px;
-}
-
-#hide {
-  position: absolute;
-  bottom: 0;
-  right: 1px;
-}
-
-#unhide {
-  position: absolute;
-  bottom: 0;
-  right: 0;
-}
-
-#hide_score {
-  position: absolute;
-  top: 5px;
-  right: 19%;
-}
-
-#unhide_score {
-  position: absolute;
-  right: 0;
-}
-
-#controlbar {
-  position: absolute;
-  height: 6vh;
-  display:flex;
-  flex-direction: column;
-  justify-content: space-around;
-  bottom: 4vh;
-  width: 99.8%;
-}
-
-#control-buttons {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-#turncontrols {
-  display: flex;
-  min-width: 220px;
-  width: 25%;
-}
-
-#filecontrols {
-  display: flex;
-  flex-direction: row;
-}
-
-#speedcontrols {
-  display: flex;
-  float: right;
-  width: 15%;
-}
-
-#end_card {
-  position: absolute;
-  top: 30vh;
-  height: 30vh;
-  width: 40%;
-  left: 30%;
-  text-align: center;
-  font-size: 50px;
-  font-family: 'Tahoma';
-  background-color: white;
-  border: 1px solid black;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-#end_card p {
-  width: 100%;
-}
-
-.file-finder {
-    height: 40px;
-    min-width: 30px;
-    font-size: 1vw;
-}
-
-#end_card button {
-  position: absolute;
-  top: -12.5px;
-  right: -12.5px;
-}
-
-#turn_progress {
-  float: right;
-  font-family: 'Tahoma';
-  font-size: 1.2vw;
-  line-height: 0;
-  margin-left: 3%;
-  align-items: center;
-  justify-content: center;
-}
-
-.speed {
-  float: left;
-  width: 50%;
-  line-height: 0;
-  font-family: 'Tahoma';
-  font-size: 1.2vw;
-}
-
-.control-button {
-  width: 2%;
-  min-width: 30px;
-}
-
-.close {
-  outline: none;
-  width: 25px;
-  height: 25px;
-  text-align: center;
-  background: inherit;
-  border-radius: 50%;
-  box-shadow: 0 0 5px rgba(0, 0, 0, 0.15);
-  border: 0;
-}
-
-#controlbar button{
-  width: 15%;
-  height: 40px;
-  min-width: 30px;
-  font-size: 1vw;
-}
-
-.controlbar p {
-  font-size: 12px;
-  width: auto;
-  float: left;
-  text-align: center;
-}
-
-
-#turn_slider{
-  width: 100%;
-}
-
-#fab {
-  display: inline-block;
-  outline: none;
-  width: 60px;
-  height: 60px;
-
-  text-align: center;
-  font-size: 32px;
-  z-index: 999;
-  background: #bf1e2e;;
-  color: #fff;
-  border-radius: 50%;
-  box-shadow: 0 0 5px rgba(0, 0, 0, 0.15);
-  border: 0;
-  overflow: hidden;
-
-  position: fixed;
-  right: 30px;
-  bottom: 120px;
-}
-
-#fab:hover {
-  transform: scale(1.05);
-}
-
-/* Slider should prob use a lib for this but I lack self respect */
-
-input[type=range] {
-  -webkit-appearance: none;
-  width: 100%;
-  margin: 13.8px 0;
-}
-
-input[type=range]:focus {
-  outline: none;
-}
-
-input[type=range]::-webkit-slider-runnable-track {
-  width: 100%;
-  height: 5px;
-  cursor: pointer;
-  background: #e6e6e6;
-  border-radius: 2.5px;
-}
-
-input[type=range]::-moz-range-track {
-  width: 100%;
-  height: 5px;
-  cursor: pointer;
-  background: #e6e6e6;
-  border-radius: 2.5px;
-}
-
-input[type=range]::-webkit-slider-thumb {
-  width: 20px;
-  height: 20px;
-  border-radius: 10px;
-  background: #1f77b4;
-  cursor: pointer;
-  -webkit-appearance: none;
-  margin-top: -7.5px;
-}
-
-input[type=range]::-moz-range-thumb {
-  width: 20px;
-  height: 20px;
-  border-radius: 10px;
-  background: #1f77b4;
-  cursor: pointer;
-  margin-top: -7.5px;
-}
-
-input[type=range]:hover::-webkit-slider-runnable-track {
-  height: 8px;
-}
-
-input[type=range]:hover::-moz-range-track {
-  height: 8px;
-}
-
-input[type=range]:hover::-webkit-slider-thumb {
-  width: 25px;
-  height: 25px;
-  margin-top: -8.5px;
-  border-radius: 12.5px;
-}
-
-input[type=range]:hover::-moz-range-thumb {
-  width: 25px;
-  height: 25px;
-  margin-top: -8.5px;
-  border-radius: 12.5px;
-}
